# logging: some simplifications/cleanups