--- a/changelog/7224.breaking.rst
+++ b/changelog/7224.breaking.rst
@@ -1,0 +1,2 @@
+The `item.catch_log_handler` and `item.catch_log_handlers` attributes, set by the
+logging plugin and never meant to be public , are no longer available.