… doc/en/deprecations.rst
… testing/deprecated_test.py
… testing/logging/test_fixture.py
… testing/logging/test_reporting.py