While looking at optimizing the logging plugin (since it shows up in profiles), and also at some issues and open PRs, I collected some simplifications and cleanups. I think they make the file a little cleaner and easier to understand and modify, and also a little faster, though this isn't the goal of this PR.

<details>
<summary>Benchmark</summary>

On this benchmark

```py
import pytest
@pytest.mark.parametrize("x", range(5000))
def test_foo(x): pass
```

Before:

    ============================ 5000 passed in 13.66s =============================
            13345882 function calls (12431103 primitive calls) in 14.054 seconds

After:

    ============================ 5000 passed in 12.53s =============================
            11271238 function calls (10716458 primitive calls) in 12.938 seconds
</details>

It is better to review each commit separately -- one of the commits reorders some methods so the PR diff will be unreadable.

Includes a few breaking changes to private attributes, and one deprecation removal (`--no-print-logs`) that was planned for 6.0.

This will cause conflicts with some open PRs. This is intentional, since I think they will be easier to evaluate after these changes (and some others). I will comment on them separately once it becomes relevant.