@bluetech 2020-05-17
Eek, looks like there is some issue on Python<=3.6. I'll take a look later.